+++ packages/std/src/exports/exports.rs
@@ -12,6 +12,9 @@ use core::{marker::PhantomData, ptr};
 
 use serde::de::DeserializeOwned;
 
+use super::imports::{ExternalApi, ExternalQuerier, ExternalStorage};
+use super::memory::{Owned, Region};
+use super::panic::install_panic_handler;
 use crate::deps::OwnedDeps;
 #[cfg(any(feature = "stargate", feature = "ibc2"))]
 use crate::ibc::IbcReceiveResponse;
@@ -635,7 +638,7 @@ where
     let info: MessageInfo = try_into_contract_result!(from_json(info));
     let msg: M = try_into_contract_result!(from_json(msg));
 
-    let mut deps = make_dependencies();
+    let mut deps = deps_from_imports();
     instantiate_fn(deps.as_mut(), env, info, msg).into()
 }
 
@@ -662,7 +665,7 @@ where
     let info: MessageInfo = try_into_contract_result!(from_json(info));
     let msg: M = try_into_contract_result!(from_json(msg));
 
-    let mut deps = make_dependencies();
+    let mut deps = deps_from_imports();
     execute_fn(deps.as_mut(), env, info, msg).into()
 }
 
@@ -685,7 +688,7 @@ where
     let env: Env = try_into_contract_result!(from_json(env));
     let msg: M = try_into_contract_result!(from_json(msg));
 
-    let mut deps = make_dependencies();
+    let mut deps = deps_from_imports();
     migrate_fn(deps.as_mut(), env, msg).into()
 }
 
@@ -712,7 +715,7 @@ where
     let msg: M = try_into_contract_result!(from_json(msg));
     let migrate_info: MigrateInfo = try_into_contract_result!(from_json(migrate_info));
 
-    let mut deps = make_dependencies();
+    let mut deps = deps_from_imports();
     migrate_with_info_fn(deps.as_mut(), env, msg, migrate_info).into()
 }
 
@@ -735,7 +738,7 @@ where
     let env: Env = try_into_contract_result!(from_json(env));
     let msg: M = try_into_contract_result!(from_json(msg));
 
-    let mut deps = make_dependencies();
+    let mut deps = deps_from_imports();
     sudo_fn(deps.as_mut(), env, msg).into()
 }
 
@@ -757,7 +760,7 @@ where
     let env: Env = try_into_contract_result!(from_json(env));
     let msg: Reply = try_into_contract_result!(from_json(msg));
 
-    let mut deps = make_dependencies();
+    let mut deps = deps_from_imports();
     reply_fn(deps.as_mut(), env, msg).into()
 }
 
@@ -779,7 +782,7 @@ where
     let env: Env = try_into_contract_result!(from_json(env));
     let msg: M = try_into_contract_result!(from_json(msg));
 
-    let deps = make_dependencies();
+    let deps = deps_from_imports();
     query_fn(deps.as_ref(), env, msg).into()
 }
 
@@ -800,7 +803,7 @@ where
     let env: Env = try_into_contract_result!(from_json(env));
     let msg: IbcChannelOpenMsg = try_into_contract_result!(from_json(msg));
 
-    let mut deps = make_dependencies();
+    let mut deps = deps_from_imports();
     contract_fn(deps.as_mut(), env, msg).into()
 }
 
@@ -823,7 +826,7 @@ where
     let env: Env = try_into_contract_result!(from_json(env));
     let msg: IbcChannelConnectMsg = try_into_contract_result!(from_json(msg));
 
-    let mut deps = make_dependencies();
+    let mut deps = deps_from_imports();
     contract_fn(deps.as_mut(), env, msg).into()
 }
 
@@ -846,7 +849,7 @@ where
     let env: Env = try_into_contract_result!(from_json(env));
     let msg: IbcChannelCloseMsg = try_into_contract_result!(from_json(msg));
 
-    let mut deps = make_dependencies();
+    let mut deps = deps_from_imports();
     contract_fn(deps.as_mut(), env, msg).into()
 }
 
@@ -869,7 +872,7 @@ where
     let env: Env = try_into_contract_result!(from_json(env));
     let msg: IbcPacketReceiveMsg = try_into_contract_result!(from_json(msg));
 
-    let mut deps = make_dependencies();
+    let mut deps = deps_from_imports();
     contract_fn(deps.as_mut(), env, msg).into()
 }
 
@@ -892,7 +895,7 @@ where
     let env: Env = try_into_contract_result!(from_json(env));
     let msg: IbcPacketAckMsg = try_into_contract_result!(from_json(msg));
 
-    let mut deps = make_dependencies();
+    let mut deps = deps_from_imports();
     contract_fn(deps.as_mut(), env, msg).into()
 }
 
@@ -915,7 +918,7 @@ where
     let env: Env = try_into_contract_result!(from_json(env));
     let msg: IbcPacketTimeoutMsg = try_into_contract_result!(from_json(msg));
 
-    let mut deps = make_dependencies();
+    let mut deps = deps_from_imports();
     contract_fn(deps.as_mut(), env, msg).into()
 }
 
@@ -937,7 +940,7 @@ where
     let env: Env = try_into_contract_result!(from_json(env));
     let msg: IbcSourceCallbackMsg = try_into_contract_result!(from_json(msg));
 
-    let mut deps = make_dependencies();
+    let mut deps = deps_from_imports();
     contract_fn(deps.as_mut(), env, msg).into()
 }
 
@@ -963,21 +966,8 @@ where
     let env: Env = try_into_contract_result!(from_json(env));
     let msg: IbcDestinationCallbackMsg = try_into_contract_result!(from_json(msg));
 
-    let mut deps = make_dependencies();
+    let mut deps = deps_from_imports();
     contract_fn(deps.as_mut(), env, msg).into()
-}
-
-/// Makes all bridges to external dependencies (i.e. Wasm imports) that are injected by the VM
-pub(crate) fn make_dependencies<Q>() -> OwnedDeps<ExternalStorage, ExternalApi, ExternalQuerier, Q>
-where
-    Q: CustomQuery,
-{
-    OwnedDeps {
-        storage: ExternalStorage::new(),
-        api: ExternalApi::new(),
-        querier: ExternalQuerier::new(),
-        custom_query_type: PhantomData,
-    }
 }
 
 #[cfg(feature = "ibc2")]
@@ -1022,7 +1012,7 @@ where
     let env: Env = try_into_contract_result!(from_json(env));
     let msg: Ibc2PacketReceiveMsg = try_into_contract_result!(from_json(msg));
 
-    let mut deps = make_dependencies();
+    let mut deps = deps_from_imports();
     contract_fn(deps.as_mut(), env, msg).into()
 }
 
@@ -1045,6 +1035,19 @@ where
     let env: Env = try_into_contract_result!(from_json(env));
     let msg: Ibc2PacketTimeoutMsg = try_into_contract_result!(from_json(msg));
 
-    let mut deps = make_dependencies();
+    let mut deps = deps_from_imports();
     contract_fn(deps.as_mut(), env, msg).into()
+}
+
+/// Makes all bridges to external dependencies (i.e. Wasm imports) that are injected by the VM
+fn deps_from_imports<Q>() -> OwnedDeps<ExternalStorage, ExternalApi, ExternalQuerier, Q>
+where
+    Q: CustomQuery,
+{
+    OwnedDeps {
+        storage: ExternalStorage::new(),
+        api: ExternalApi::new(),
+        querier: ExternalQuerier::new(),
+        custom_query_type: PhantomData,
+    }
 }