+++ packages/std/src/lib.rs
@@ -33,7 +33,6 @@ mod metadata;
 mod msgpack;
 mod never;
 mod pagination;
-mod panic;
 mod query;
 mod results;
 mod sections;
@@ -97,9 +96,8 @@ pub use crate::query::{
     CustomQuery, DecCoin, Delegation, DelegationResponse, DelegationRewardsResponse,
     DelegationTotalRewardsResponse, DelegatorReward, DelegatorValidatorsResponse,
     DelegatorWithdrawAddressResponse, DenomMetadataResponse, DistributionQuery,
-    FeeEnabledChannelResponse, FullDelegation, GrpcQuery, IbcQuery, ListChannelsResponse,
-    PortIdResponse, QueryRequest, StakingQuery, SupplyResponse, Validator, ValidatorResponse,
-    WasmQuery,
+    FeeEnabledChannelResponse, FullDelegation, GrpcQuery, IbcQuery, PortIdResponse, QueryRequest,
+    StakingQuery, SupplyResponse, Validator, ValidatorResponse, WasmQuery,
 };
 #[cfg(all(feature = "stargate", feature = "cosmwasm_1_2"))]
 pub use crate::results::WeightedVoteOption;
@@ -118,30 +116,27 @@ pub use crate::timestamp::Timestamp;
 pub use crate::traits::{Api, HashFunction, Querier, QuerierResult, QuerierWrapper, Storage};
 pub use crate::types::{BlockInfo, ContractInfo, Env, MessageInfo, MigrateInfo, TransactionInfo};
 
-// Exposed in wasm build only
-#[cfg(target_arch = "wasm32")]
-mod exports;
-#[cfg(target_arch = "wasm32")]
-mod imports;
-#[cfg(target_arch = "wasm32")]
-mod memory; // Used by exports and imports only. This assumes pointers are 32 bit long, which makes it untestable on dev machines.
+//
+// Exports
+//
 
-#[cfg(all(feature = "cosmwasm_2_2", target_arch = "wasm32"))]
+#[cfg(all(feature = "exports", target_arch = "wasm32"))]
+mod exports;
+
+#[cfg(all(feature = "exports", target_arch = "wasm32", feature = "cosmwasm_2_2"))]
 pub use crate::exports::do_migrate_with_info;
-#[cfg(target_arch = "wasm32")]
+#[cfg(all(feature = "exports", target_arch = "wasm32"))]
 pub use crate::exports::{
     do_execute, do_ibc_destination_callback, do_ibc_source_callback, do_instantiate, do_migrate,
     do_query, do_reply, do_sudo,
 };
-#[cfg(all(feature = "ibc2", target_arch = "wasm32"))]
+#[cfg(all(feature = "exports", target_arch = "wasm32", feature = "ibc2"))]
 pub use crate::exports::{do_ibc2_packet_ack, do_ibc2_packet_receive, do_ibc2_packet_timeout};
-#[cfg(all(feature = "stargate", target_arch = "wasm32"))]
+#[cfg(all(feature = "exports", target_arch = "wasm32", feature = "stargate"))]
 pub use crate::exports::{
     do_ibc_channel_close, do_ibc_channel_connect, do_ibc_channel_open, do_ibc_packet_ack,
     do_ibc_packet_receive, do_ibc_packet_timeout,
 };
-#[cfg(target_arch = "wasm32")]
-pub use crate::imports::{ExternalApi, ExternalQuerier, ExternalStorage};
 
 /// Exposed for testing only
 /// Both unit tests and integration tests are compiled to native code, so everything in here does not need to compile to Wasm.