+++ packages/std/src/testing/mod.rs
@@ -19,8 +19,8 @@ pub use mock::DistributionQuerier;
 pub use mock::StakingQuerier;
 pub use mock::{
     mock_dependencies, mock_dependencies_with_balance, mock_dependencies_with_balances, mock_env,
-    mock_wasmd_attr, BankQuerier, MockApi, MockQuerier, MockQuerierCustomHandlerResult,
-    MOCK_CONTRACT_ADDR,
+    mock_wasmd_attr, BankQuerier, Envs, EnvsOptions, MockApi, MockQuerier,
+    MockQuerierCustomHandlerResult, MOCK_CONTRACT_ADDR,
 };
 #[cfg(feature = "ibc2")]
 pub use mock::{mock_ibc2_packet_recv, mock_ibc2_packet_timeout, mock_ibc2_packet_ack};