+++ packages/vm/src/static_analysis.rs
@@ -64,9 +64,6 @@ pub const REQUIRED_IBC_EXPORTS: &[Entrypoint] = &[
     Entrypoint::IbcPacketTimeout,
 ];
 
-pub const REQUIRED_IBC2_EXPORT: &[Entrypoint] =
-    &[Entrypoint::Ibc2PacketReceive, Entrypoint::Ibc2PacketTimeout];
-
 /// A trait that allows accessing shared functionality of `parity_wasm::elements::Module`
 /// and `wasmer::Module` in a shared fashion.
 pub trait ExportInfo {